backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2 ; python_version >= '3.7'
django==4.1.2
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-stubs==1.12.0
django-stubs-ext==0.5.0 ; python_version >= '3.6'
djangorestframework==3.14.0
factory-boy==3.2.1
faker==15.1.1 ; python_version >= '3.7'
markdown==3.4.1
mypy==0.982 ; python_version >= '3.7'
mypy-extensions==0.4.3
psycopg2==2.9.4
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.5
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3 ; python_version >= '3.5'
tomli==2.0.1 ; python_version >= '3.7'
types-pytz==2022.5.0.0
types-pyyaml==6.0.12
typing-extensions==4.4.0 ; python_version >= '3.7'
